+++ src/databaseTest/java/net/sf/jabref/shared/SynchronizationTestSimulator.java
@@ -6,11 +6,9 @@ import java.util.Collection;
 
 import net.sf.jabref.BibDatabaseContext;
 import net.sf.jabref.Defaults;
-import net.sf.jabref.Globals;
 import net.sf.jabref.model.database.BibDatabaseMode;
 import net.sf.jabref.model.database.DatabaseLocation;
 import net.sf.jabref.model.entry.BibEntry;
-import net.sf.jabref.preferences.JabRefPreferences;
 import net.sf.jabref.shared.exception.DatabaseNotSupportedException;
 
 import org.junit.After;
@@ -43,12 +41,10 @@ public class SynchronizationTestSimulator {
             connection = TestConnector.getTestConnection(dbmsType);
         }
 
-        Globals.prefs = JabRefPreferences.getInstance();
-
-        clientContextA = new BibDatabaseContext(new Defaults(BibDatabaseMode.BIBTEX), DatabaseLocation.SHARED);
+        clientContextA = new BibDatabaseContext(new Defaults(BibDatabaseMode.BIBTEX), DatabaseLocation.SHARED, ", ");
         clientContextA.getDBSynchronizer().openSharedDatabase(connection, dbmsType, "A");
 
-        clientContextB = new BibDatabaseContext(new Defaults(BibDatabaseMode.BIBTEX), DatabaseLocation.SHARED);
+        clientContextB = new BibDatabaseContext(new Defaults(BibDatabaseMode.BIBTEX), DatabaseLocation.SHARED, ", ");
         clientContextB.getDBSynchronizer().openSharedDatabase(connection, dbmsType, "B");
         eventListenerB = new SynchronizationTestEventListener();
         clientContextB.getDBSynchronizer().registerListener(eventListenerB);