+++ src/main/java/net/sf/jabref/gui/importer/ImportInspectionDialog.java
@@ -458,7 +458,7 @@ public class ImportInspectionDialog extends JDialog implements ImportInspector, 
 
             entry.setId(IdGenerator.next());
             // Add the entry to the database we are working with:
-            database.insertEntry(entry);
+            database.insertEntryWithDuplicationCheck(entry);
 
             // Generate a unique key:
             BibtexKeyPatternUtil.makeLabel(localMetaData, database, entry,
@@ -501,7 +501,7 @@ public class ImportInspectionDialog extends JDialog implements ImportInspector, 
             for (BibEntry entry : entries) {
 
                 entry.setId(IdGenerator.next());
-                database.insertEntry(entry);
+                database.insertEntryWithDuplicationCheck(entry);
 
                 BibtexKeyPatternUtil.makeLabel(localMetaData, database, entry,
                         Globals.prefs.getBibtexKeyPatternPreferences());
@@ -735,7 +735,7 @@ public class ImportInspectionDialog extends JDialog implements ImportInspector, 
                 }
 
                 entry.setId(IdGenerator.next());
-                panel.getDatabase().insertEntry(entry);
+                panel.getDatabase().insertEntryWithDuplicationCheck(entry);
                 ce.addEdit(new UndoableInsertEntry(panel.getDatabase(), entry, panel));
 
             }