+++ src/main/java/net/sf/jabref/logic/integrity/IntegrityCheck.java
@@ -306,9 +306,17 @@ public class IntegrityCheck {
     private static class YearChecker implements Checker {
 
         private static final Predicate<String> CONTAINS_FOUR_DIGIT = Pattern.compile("([^0-9]|^)[0-9]{4}([^0-9]|$)").asPredicate();
+        private static final Predicate<String> ENDS_WITH_FOUR_DIGIT = Pattern.compile("[0-9]{4}$")
+                .asPredicate();
+        private static final String PUNCTUATION_MARKS = "[(){},.;!?<>%&$]";
+
 
         /**
-         * Checks, if the number String contains a four digit year
+         * Checks, if the number String contains a four digit year and ends with it.
+         * Official bibtex spec:
+         * Generally it should consist of four numerals, such as 1984, although the standard styles
+         * can handle any year whose last four nonpunctuation characters are numerals, such as ‘(about 1984)’.
+         * Source: http://ftp.fernuni-hagen.de/ftp-dir/pub/mirrors/www.ctan.org/biblio/bibtex/base/btxdoc.pdf
          */
         @Override
         public List<IntegrityMessage> check(BibEntry entry) {
@@ -319,6 +327,12 @@ public class IntegrityCheck {
 
             if (!CONTAINS_FOUR_DIGIT.test(value.get().trim())) {
                 return Collections.singletonList(new IntegrityMessage(Localization.lang("should contain a four digit number"), entry, FieldName.YEAR));
+            }
+
+            if (!ENDS_WITH_FOUR_DIGIT.test(value.get().replaceAll(PUNCTUATION_MARKS, ""))) {
+                return Collections.singletonList(new IntegrityMessage(
+                        Localization.lang("last four nonpunctuation characters should be numerals"), entry,
+                        FieldName.YEAR));
             }
 
             return Collections.emptyList();