+++ src/main/java/net/sf/jabref/model/database/BibDatabase.java
@@ -13,13 +13,11 @@ import java.util.Set;
 import java.util.TreeSet;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 
 import net.sf.jabref.event.source.EntryEventSource;
-import net.sf.jabref.model.EntryTypes;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.BibtexString;
-import net.sf.jabref.model.entry.EntryType;
-import net.sf.jabref.model.entry.EntryUtil;
 import net.sf.jabref.model.entry.FieldName;
 import net.sf.jabref.model.entry.InternalBibtexFields;
 import net.sf.jabref.model.entry.MonthUtil;
@@ -37,7 +35,6 @@ import org.apache.commons.logging.LogFactory;
  * A bibliography database.
  */
 public class BibDatabase {
-
     private static final Log LOGGER = LogFactory.getLog(BibDatabase.class);
 
     /**
@@ -59,7 +56,6 @@ public class BibDatabase {
      * contains all entry.getID() of the current database
      */
     private final Set<String> internalIDs = new HashSet<>();
-
 
     private final EventBus eventBus = new EventBus();
 
@@ -101,22 +97,18 @@ public class BibDatabase {
         return Collections.unmodifiableList(entries);
     }
 
+    /**
+     * Returns a set of Strings, that contains all field names that are visible. This means that the fields
+     * are not internal fields. Internal fields are fields, that are starting with "_".
+     *
+     * @return set of fieldnames, that are visible
+     */
     public Set<String> getAllVisibleFields() {
         Set<String> allFields = new TreeSet<>();
         for (BibEntry e : getEntries()) {
             allFields.addAll(e.getFieldNames());
         }
-        Set<String> toberemoved = new TreeSet<>();
-        for (String field : allFields) {
-            if (InternalBibtexFields.isInternalField(field)) {
-                toberemoved.add(field);
-            }
-        }
-
-        for (String field : toberemoved) {
-            allFields.remove(field);
-        }
-        return allFields;
+        return allFields.stream().filter(field -> !InternalBibtexFields.isInternalField(field)).collect(Collectors.toSet());
     }
 
     /**
@@ -131,6 +123,13 @@ public class BibDatabase {
         return Optional.empty();
     }
 
+    /**
+     * Collects entries having the specified BibTeX key and returns these entries as list.
+     * The order of the entries is the order they appear in the database.
+     *
+     * @param key
+     * @return list of entries that contains the given key
+     */
     public synchronized List<BibEntry> getEntriesByKey(String key) {
         List<BibEntry> result = new ArrayList<>();
 
@@ -152,8 +151,14 @@ public class BibDatabase {
      * @return false if the insert was done without a duplicate warning
      * @throws KeyCollisionException thrown if the entry id ({@link BibEntry#getId()}) is already  present in the database
      */
-    public synchronized boolean insertEntry(BibEntry entry) throws KeyCollisionException {
-        return insertEntry(entry, EntryEventSource.LOCAL);
+    @Deprecated // use insertEntry and DuplicationChecker separately
+    public synchronized boolean insertEntryWithDuplicationCheck(BibEntry entry) throws KeyCollisionException {
+        insertEntry(entry);
+        return duplicationChecker.checkForDuplicateKeyAndAdd(null, entry.getCiteKey());
+    }
+
+    public synchronized void insertEntry(BibEntry entry) {
+        insertEntry(entry, EntryEventSource.LOCAL);
     }
 
     /**
@@ -162,10 +167,8 @@ public class BibDatabase {
      *
      * @param entry BibEntry to insert
      * @param eventSource Source the event is sent from
-     * @return false if the insert was done without a duplicate warning
      */
-    public synchronized boolean insertEntry(BibEntry entry, EntryEventSource eventSource)
-            throws KeyCollisionException {
+    public synchronized void insertEntry(BibEntry entry, EntryEventSource eventSource) throws KeyCollisionException {
         Objects.requireNonNull(entry);
 
         String id = entry.getId();
@@ -178,7 +181,6 @@ public class BibDatabase {
         entry.registerListener(this);
 
         eventBus.post(new EntryAddedEvent(entry, eventSource));
-        return duplicationChecker.checkForDuplicateKeyAndAdd(null, entry.getCiteKeyOptional().orElse(null));
     }
 
     /**
@@ -193,13 +195,14 @@ public class BibDatabase {
     /**
      * Removes the given entry.
      * The Entry is removed based on the id {@link BibEntry#id}
+     *
      * @param toBeDeleted Entry to delete
      * @param eventSource Source the event is sent from
      */
     public synchronized void removeEntry(BibEntry toBeDeleted, EntryEventSource eventSource) {
         Objects.requireNonNull(toBeDeleted);
 
-        boolean anyRemoved =  entries.removeIf(entry -> entry.getId().equals(toBeDeleted.getId()));
+        boolean anyRemoved = entries.removeIf(entry -> entry.getId().equals(toBeDeleted.getId()));
         if (anyRemoved) {
             internalIDs.remove(toBeDeleted.getId());
             toBeDeleted.getCiteKeyOptional().ifPresent(duplicationChecker::removeKeyFromSet);
@@ -211,10 +214,17 @@ public class BibDatabase {
         return duplicationChecker.getNumberOfKeyOccurrences(key);
     }
 
+    /**
+     * Sets the given key to the given entry.
+     * If the key is null, the entry field will be cleared.
+     *
+     * @return true, if the entry contains the key, false if not
+     */
+    @Deprecated // use BibEntry.setCiteKey (and DuplicationChecker)
     public synchronized boolean setCiteKeyForEntry(BibEntry entry, String key) {
-        String oldKey = entry.getCiteKeyOptional().orElse(null);
+        String oldKey = entry.getCiteKey();
         if (key == null) {
-            entry.clearField(BibEntry.KEY_FIELD);
+            entry.clearCiteKey();
         } else {
             entry.setCiteKey(key);
         }
@@ -238,14 +248,13 @@ public class BibDatabase {
     /**
      * Inserts a Bibtex String.
      */
-    public synchronized void addString(BibtexString string)
-            throws KeyCollisionException {
+    public synchronized void addString(BibtexString string) throws KeyCollisionException {
         if (hasStringLabel(string.getName())) {
             throw new KeyCollisionException("A string with that label already exists");
         }
 
         if (bibtexStrings.containsKey(string.getId())) {
-            throw new KeyCollisionException("Duplicate BibTeXString id.");
+            throw new KeyCollisionException("Duplicate BibTeX string id.");
         }
 
         bibtexStrings.put(string.getId(), string);
@@ -479,61 +488,6 @@ public class BibDatabase {
         return res;
     }
 
-
-
-    /**
-     * Returns the text stored in the given field of the given bibtex entry
-     * which belongs to the given database.
-     * <p>
-     * If a database is given, this function will try to resolve any string
-     * references in the field-value.
-     * Also, if a database is given, this function will try to find values for
-     * unset fields in the entry linked by the "crossref" field, if any.
-     *
-     * @param field    The field to return the value of.
-     * @param entry    The bibtex entry which contains the field.
-     * @param database maybenull
-     *                 The database of the bibtex entry.
-     * @return The resolved field value or null if not found.
-     */
-    public static Optional<String> getResolvedField(String field, BibEntry entry, BibDatabase database) {
-        Objects.requireNonNull(entry, "entry cannot be null");
-
-        if (BibEntry.TYPE_HEADER.equals(field) || BibEntry.OBSOLETE_TYPE_HEADER.equals(field)) {
-            Optional<EntryType> entryType = EntryTypes.getType(entry.getType(), BibDatabaseMode.BIBLATEX);
-            if (entryType.isPresent()) {
-                return Optional.of(entryType.get().getName());
-            } else {
-                return Optional.of(EntryUtil.capitalizeFirst(entry.getType()));
-            }
-        }
-
-        if (BibEntry.KEY_FIELD.equals(field)) {
-            return entry.getCiteKeyOptional();
-        }
-
-        // TODO: Changed this to also consider alias fields, which is the expected
-        // behavior for the preview layout and for the check whatever all fields are present.
-        // But there might be unwanted side-effects?!
-        Optional<String> result = entry.getFieldOrAlias(field);
-
-        // If this field is not set, and the entry has a crossref, try to look up the
-        // field in the referred entry: Do not do this for the bibtex key.
-        if (!result.isPresent() && (database != null)) {
-            Optional<String> crossrefKey = entry.getField(FieldName.CROSSREF);
-            if (crossrefKey.isPresent() && !crossrefKey.get().isEmpty()) {
-                Optional<BibEntry> referred = database.getEntryByKey(crossrefKey.get());
-                if (referred.isPresent()) {
-                    // Ok, we found the referred entry. Get the field value from that
-                    // entry. If it is unset there, too, stop looking:
-                    result = referred.get().getFieldOrAlias(field);
-                }
-            }
-        }
-
-        return result.map(resultText -> BibDatabase.getText(resultText, database));
-    }
-
     /**
      * Returns a text with references resolved according to an optionally given database.
      *
@@ -541,6 +495,7 @@ public class BibDatabase {
      * @param database  maybenull The database to use for resolving the text.
      * @return The resolved text or the original text if either the text or the database are null
      */
+    @Deprecated // use BibDatabase.resolveForStrings
     public static String getText(String toResolve, BibDatabase database) {
         if ((toResolve != null) && (database != null)) {
             return database.resolveForStrings(toResolve);
@@ -581,5 +536,9 @@ public class BibDatabase {
     @Subscribe
     private void relayEntryChangeEvent(FieldChangedEvent event) {
         eventBus.post(event);
+    }
+
+    public Optional<BibEntry> getReferencedEntry(BibEntry entry) {
+        return entry.getField(FieldName.CROSSREF).flatMap(this::getEntryByKey);
     }
 }